+++ CHANGELOG.md
@@ -3107,3 +3107,4 @@
 - [2026-07-24 12:12] perf(test): apply RepaintBoundary on animated test widgets
 - [2026-07-24 12:51] feat(ui): implement responsive ui view component
 - [2026-07-24 13:39] feat(auth): implement responsive auth view component
+- [2026-07-24 13:49] test(state): add unit tests for state business logic